requirements.txt
zenml[server]
mlflow
tensorflow
scikit-learn
matplotlib
seaborn
Pillow
python-dotenv
dvc
tensorboard
boto3
